+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './k-number';
 
 export * from './constants';
-export * from './utils';